Lift compute_world_directions + filter_tracks into utils.tracks
Addresses Akshay's review comment about duplicated code in
global_positioner.py:
- compute_world_directions: same per-measurement logic as
  averaging_1dsfm._get_landmark_directions (calibrate → unit ray →
  rotate to world frame). Now lives in gtsfm.utils.tracks and is
  shared by both global positioner and 1DSfM translation averaging.
- filter_tracks_by_measurements: track filter (keep tracks with
  enough valid-camera measurements, longest first, optional cap).
  Lifted to gtsfm.utils.tracks for reuse.

1DSfM's _get_landmark_directions is now a thin adapter that calls
the shared util and re-shapes the List[Tuple] output into the
existing Dict[(track, cam), Unit3] format. Behavior is identical;
the asserts on None wRi/intrinsics relax to silent skips, which is
strictly safer (the upstream caller guarantees validity, the asserts
just panic on edge cases).

Also in this commit:
- Replace sift_global_positioner.yaml with the production MegaLoc-
  based config so the team can run experiments without juggling
  two near-identical config files.
- Regenerate uv.lock for the gtsam-develop pin bump from the prior
  commit (CI was failing because --locked saw drift).
