+++ gtsfm/configs/sift_global_positioner.yaml
@@ -1,13 +1,5 @@
-# SIFT front-end with Global Positioner (replaces translation averaging + data association).
-#
-# Uses the GLOMAP-inspired BATA bearing factor graph to jointly estimate
-# camera positions and 3D points, feeding directly into global BA.
-#
-# Pipeline:  rot_avg → GlobalPositioner → BA
-# (instead of: rot_avg → trans_avg → data_assoc → BA)
-#
-# Front-end matches unified.yaml exactly for fair A/B comparison with CI benchmarks.
-# Only the multiview_optimizer differs: global_positioner is added.
+# MegaLoc + SIFT + Global Positioner with Single partition.
+# Based on unified_megaloc.yaml with SIFT detector/matcher swap + GP.
 
 # @package _global_
 _target_: gtsfm.scene_optimizer.SceneOptimizer
@@ -20,27 +12,29 @@ image_pairs_generator:
   global_descriptor:
     _target_: gtsfm.frontend.cacher.global_descriptor_cacher.GlobalDescriptorCacher
     global_descriptor_obj:
-      _target_: gtsfm.frontend.global_descriptor.NetVLAD
+      _target_: gtsfm.frontend.global_descriptor.MegaLoc
   retriever:
-    _target_: gtsfm.retriever.JointSimilaritySequential
-    num_matched: 5
-    min_score: 0.3
-    max_frame_lookahead: 15
-  batch_size: 2
+    _target_: gtsfm.retriever.Similarity
+    num_matched: 100
+    min_score: 0.15
+    use_diffusion_preprocessing: false
+    use_mst_augmentation: false
+  batch_size: 16
 
 graph_partitioner:
   _target_: gtsfm.graph_partitioner.Single
 
 cluster_optimizer:
   _target_: gtsfm.cluster_optimizer.Multiview
+
   correspondence_generator:
     _target_: gtsfm.frontend.correspondence_generator.det_desc_correspondence_generator.DetDescCorrespondenceGenerator
 
     detector_descriptor:
       _target_: gtsfm.frontend.cacher.detector_descriptor_cacher.DetectorDescriptorCacher
       detector_descriptor_obj:
-        _target_: gtsfm.frontend.detector_descriptor.sift.SIFTDetectorDescriptor
-        max_keypoints: 5000
+        _target_: gtsfm.frontend.detector_descriptor.colmap_sift.ColmapSIFTDetectorDescriptor
+        max_keypoints: 8192
 
     matcher:
       _target_: gtsfm.frontend.cacher.matcher_cacher.MatcherCacher
@@ -53,14 +47,13 @@ cluster_optimizer:
     two_view_estimator_obj:
       _target_: gtsfm.two_view_estimator.TwoViewEstimator
       bundle_adjust_2view: True
-      eval_threshold_px: 4 # in px
+      eval_threshold_px: 4
       ba_reproj_error_thresholds: [0.5]
       bundle_adjust_2view_maxiters: 100
 
       verifier:
-        _target_: gtsfm.frontend.verifier.ransac.Ransac
-        use_intrinsics_in_verification: True
-        estimation_threshold_px: 4 # for H/E/F estimators
+        _target_: gtsfm.frontend.verifier.poselib_verifier.PoseLibVerifier
+        estimation_threshold_px: 2
 
       triangulation_options:
         _target_: gtsfm.data_association.point3d_initializer.TriangulationOptions
@@ -70,26 +63,24 @@ cluster_optimizer:
 
       inlier_support_processor:
         _target_: gtsfm.two_view_estimator.InlierSupportProcessor
-        min_num_inliers_est_model: 15
-        min_inlier_ratio_est_model: 0.1
+        min_num_inliers_est_model: 30
+        min_inlier_ratio_est_model: 0.15
   save_gtsfm_data: True
   save_3d_viz: False
   save_two_view_viz: False
-  pose_angular_error_thresh: 5 # degrees
+  pose_angular_error_thresh: 5
   multiview_optimizer:
     _target_: gtsfm.multi_view_optimizer.MultiViewOptimizer
 
-    # Matches unified.yaml exactly
     view_graph_estimator:
       _target_: gtsfm.view_graph_estimator.cycle_consistent_rotation_estimator.CycleConsistentRotationViewGraphEstimator
       edge_error_aggregation_criterion: MIN_EDGE_ERROR
+      error_threshold: 7.0
 
     rot_avg_module:
       _target_: gtsfm.averaging.rotation.shonan.ShonanRotationAveraging
       weight_by_inliers: True
 
-    # trans_avg_module and data_association_module are required for
-    # backwards compatibility but are unused when global_positioner is set.
     trans_avg_module:
       _target_: gtsfm.averaging.translation.averaging_1dsfm.TranslationAveraging1DSFM
       robust_measurement_noise: True
@@ -110,16 +101,17 @@ cluster_optimizer:
         max_num_hypotheses: 100
       save_track_patches_viz: False
 
-    # Global Positioner: replaces trans_avg + data_assoc with joint optimization.
-    # Uses GTSAM's native BilinearAngleTranslationFactor (C++) for full speed.
     global_positioner:
       _target_: gtsfm.global_positioner.global_positioner.GlobalPositioner
       noise_sigma: 0.01
       huber_loss_scale: 0.1
-      max_iterations: 30
+      max_iterations: 50
       min_track_measurements: 3
-      max_reproj_error: 5.0
+      max_reproj_error: 10.0
       save_iteration_visualization: True
+
+    run_view_graph_calibration: true
+    rotation_outlier_threshold_deg: 0
 
     bundle_adjustment_module:
       _target_: gtsfm.bundle.bundle_adjustment.BundleAdjustmentOptimizer
@@ -127,14 +119,10 @@ cluster_optimizer:
       robust_ba_mode: "HUBER"
       shared_calib: False
       cam_pose3_prior_noise_sigma: 0.1
-      calibration_prior_focal_sigma: 20.0
+      calibration_prior_focal_sigma: 200.0
       measurement_noise_sigma: 1.0
+      use_calibration_prior: true
 
-  # comment out to not run
-  dense_multiview_optimizer:
-    _target_: gtsfm.densify.mvs_patchmatchnet.MVSPatchmatchNet
+  dense_multiview_optimizer: null
+  gaussian_splatting_optimizer: null
 
-  gaussian_splatting_optimizer:
-    _target_: gtsfm.splat.gaussian_splatting.GaussianSplatting
-    cfg:
-      _target_: gtsfm.splat.gaussian_splatting.Config
